requirements.txt
asgiref==3.2.3
astroid==2.3.2
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
boto3==1.10.49
botocore==1.13.49
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-cities==0.5.0.6
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.2.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-mssql==1.8
django-payu==0.5
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-storages==1.8
django-stubs==1.4.0
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
docutils==0.15.2
entrypoints==0.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
mypy==0.761
mypy-extensions==0.4.3
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
oauthlib==3.0.1
pandocfilters==1.4.2
parso==0.5.2
pbr==5.4.4
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.3
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.2.4
stevedore==1.31.0
swapper==1.1.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.24.2
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==5.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
